requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.3
bs4==0.0.1
cachetools==4.2.1
cchardet==2.1.7
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
contextvars==2.4
cycler==0.10.0
# dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
dnspython==1.16.0
entrypoints==0.3
fake-useragent==0.1.11
fastprogress==1.0.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googletrans==3.1.0a0
grpcio==1.35.0
h11==0.9.0
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
immutables==0.14
importlib-metadata==3.4.0
ipykernel==5.4.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
jieba==0.42.1
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
keras-bert==0.86.0
keras-embed-sim==0.8.0
keras-layer-normalization==0.14.0
keras-multi-head==0.27.0
keras-pos-embd==0.11.0
keras-position-wise-feed-forward==0.6.0
Keras-Preprocessing==1.1.2
keras-self-attention==0.46.0
keras-transformer==0.38.0
kiwisolver==1.3.1
ktrain==0.25.4
langdetect==1.0.8
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
nltk==3.5
notebook==6.2.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
pymongo==3.11.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
#pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.2
qtconsole==5.0.2
QtPy==1.9.0
rake-nltk==1.0.4
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.7
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
Send2Trash==1.5.0
sentencepiece==0.1.91
seqeval==0.0.19
six==1.15.0
sklearn==0.0
sniffio==1.2.0
soupsieve==2.1
syntok==1.3.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-gpu==2.3.1
tensorflow-gpu-estimator==2.3.0
tensorflow-hub==0.10.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
testresources==2.0.1
threadpoolctl==2.1.0
tokenizers==0.9.3
tornado==6.1
tqdm==4.56.0
traitlets==4.3.3
transformers==3.5.1
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
Whoosh==2.7.4
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.0
